requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-surprise>=1.1.3
tqdm>=4.66.0
